feat: Add feature flag support in the fib service (#142)
## This PR

- adds an experimental OpenFeature OTel propagator
- adds OpenFeature to the fib service
- adds dynamic provider switcher to fib service

### Related Issues

Fixes #123 

### Notes

The OpenFeature OTel propagator automatically sets evaluation context to
an HTTP header on each request. This data is then extracted by the fib
service and included in the request scoped context. The technique is
experimental but it could prove extremely valuable in microservice
architectures where important evaluation context is omitted.

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>